+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.0.3 - WIP
 - Make `pan_length` first parameter of `pan_range`.
 - China UnionPay, China T-Union, Diners Club, Diners Club International,
-- JCB, Maestro UK, Visa Electron
+- JCB, Maestro, Maestro UK, Visa Electron
 
 ## v0.0.2 - 2023-08-03
 - Initial release.